backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
openai
pydantic
pydantic-settings
python-dotenv
aiofiles
